Extend and rename test
